Extract Estimator and DagOperations from MultiParentCasper
